Fix test failure with ReinterpretArray
next(x, endof(x)) is not generally valid with AbstractVectors even though
it works for Vector. This fixes a test failure on Julia 0.7.

The UTF32String(::Vector{Char}) constructor added a \0 char at the end
of the string because it used reinterpret(UInt32, ...), which did not
dispatch to the inner constructor directly as it used to do on 0.6.